CLOUD-45130 fix caching in aws client
